refactor: flytta gallringsregel-whitelist till core/common (CodeRabbit R1) + javadoc
- Flyttade allowedMetadataConditionFields() från plugin-utilen till
  org.roda.core.common.DisposalRuleConditionFields. roda-ui anropar nu den delade
  core/common-klassen i stället för en plugin-implementationsklass, vilket
  respekterar lagergränsen (CodeRabbit R1). Både apply-jobbet och
  DisposalRuleService använder samma källa.
- La till Javadoc på publika ApplyDisposalRulesPluginUtils.applyRule (R6).

Den tidigare docs/issues-MD022-noten är inaktuell (filen är borttagen ur PR:n).
